requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10      
aiosignal==1.3.1      
annotated-types==0.7.0
anyio==4.6.2
attrs==24.2.0
azure-functions==1.21.3
certifi==2024.8.30
colorama==0.4.6
discord.py==2.4.0
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
multidict==6.1.0
openai==1.51.2
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
yarl==1.15.2